+++ R/performance_rmse.R
@@ -6,7 +6,7 @@
 #'
 #' @param model A model.
 #' @param normalized Logical, use `TRUE` if normalized rmse should be returned.
-#' @inheritParams model_performance.lm
+#' @inheritParams icc
 #'
 #' @details The RMSE is the square root of the variance of the residuals and indicates
 #'   the absolute fit of the model to the data (difference between observed data
@@ -30,33 +30,138 @@
 #' # normalized RMSE
 #' performance_rmse(m, normalized = TRUE)
 #' @export
-performance_rmse <- function(model, normalized = FALSE, verbose = TRUE) {
+performance_rmse <- function(model,
+                             normalized = FALSE,
+                             ci = NULL,
+                             iterations = 100,
+                             ci_method = NULL,
+                             verbose = TRUE,
+                             ...) {
   tryCatch(
     {
-      # compute rmse
-      rmse_val <- sqrt(performance_mse(model, verbose = verbose))
-
-      # if normalized, divide by range of response
-      if (normalized) {
-        # get response
-        resp <- datawizard::to_numeric(insight::get_response(model, verbose = FALSE), dummy_factors = FALSE, preserve_levels = TRUE)
-        # compute rmse, normalized
-        rmse_val <- rmse_val / (max(resp, na.rm = TRUE) - min(resp, na.rm = TRUE))
+      out <- .calculate_rmse(model, normalized, verbose)
+      # check if CIs are requested, and compute CIs
+      if (!is.null(ci) && !is.na(ci)) {
+        # analytical CI?
+        if (identical(ci_method, "analytical")) {
+          out <- .analytical_rmse_ci(out, model, ci)
+        } else {
+          # bootstrapped CI
+          result <- .bootstrap_rmse(model, iterations, normalized, ci_method, ...)
+          # CI for RMSE
+          rmse_ci <- as.vector(result$t[, 1])
+          rmse_ci <- rmse_ci[!is.na(rmse_ci)]
+          # validation check
+          if (length(rmse_ci) > 0) {
+            rmse_ci <- bayestestR::eti(rmse_ci, ci = ci)
+            out <- cbind(data.frame(RMSE = out), rmse_ci)
+            class(out) <- c("performance_rmse", "data.frame")
+          } else {
+            insight::format_warning("Could not compute confidence intervals for RMSE.")
+          }
+        }
       }
-
-      rmse_val
     },
     error = function(e) {
       if (inherits(e, c("simpleError", "error")) && verbose) {
         insight::print_color(e$message, "red")
         cat("\n")
       }
-      NA
+      out <- NA
     }
   )
+
+  out
 }
 
 
 #' @rdname performance_rmse
 #' @export
 rmse <- performance_rmse
+
+
+# methods ---------------------------------------------------------------------
+
+#' @export
+format.performance_rmse <- function(x, ...) {
+  insight::format_table(x, ...)
+}
+
+#' @export
+print.performance_rmse <- function(x, ...) {
+  cat(insight::export_table(format(x, ...), ...))
+}
+
+
+# helper function to compute RMSE ----------------------------------------------
+
+.calculate_rmse <- function(model, normalized = FALSE, verbose = FALSE, ...) {
+  # compute rmse
+  rmse_val <- sqrt(performance_mse(model, verbose = verbose))
+
+  # if normalized, divide by range of response
+  if (normalized) {
+    # get response
+    resp <- datawizard::to_numeric(
+      insight::get_response(model, verbose = FALSE),
+      dummy_factors = FALSE,
+      preserve_levels = TRUE
+    )
+    # compute rmse, normalized
+    rmse_val <- rmse_val / (max(resp, na.rm = TRUE) - min(resp, na.rm = TRUE))
+  }
+
+  rmse_val
+}
+
+
+# analytical CIs --------------------------------------------------------------
+
+.analytical_rmse_ci <- function(out, model, ci, ...) {
+  s <- insight::get_sigma(model, ci = ci, verbose = FALSE)
+  n <- insight::n_obs(model)
+  conf_ints <- c(attr(s, "CI_low"), attr(s, "CI_high")) * ((n - 1) / n)
+  out <- data.frame(
+    RMSE = out,
+    CI = ci,
+    CI_low = conf_ints[1],
+    CI_high = conf_ints[2]
+  )
+  class(out) <- c("performance_rmse", "data.frame")
+  out
+}
+
+
+# bootstrapping CIs -----------------------------------------------------------
+
+.boot_calculate_rmse <- function(data, indices, model, normalized, ...) {
+  d <- data[indices, ] # allows boot to select sample
+  fit <- suppressWarnings(suppressMessages(stats::update(model, data = d)))
+  .calculate_rmse(model = fit, normalized = normalized)
+}
+
+.bootstrap_rmse <- function(model, iterations = 100, normalized = FALSE, ci_method = NULL, ...) {
+  if (inherits(model, c("merMod", "lmerMod", "glmmTMB")) && !identical(ci_method, "boot")) {
+    # cannot pass argument "normalized" to "lme4::bootMer()"
+    if (isTRUE(normalized)) {
+      insight::format_error("Normalized RMSE cannot be used with confidence intervals. Please use `ci_method = \"boot\"`.") # nolint
+    }
+    result <- .do_lme4_bootmer(
+      model,
+      .calculate_rmse,
+      iterations,
+      dots = list(...)
+    )
+  } else {
+    insight::check_if_installed("boot")
+    result <- boot::boot(
+      data = insight::get_data(model, verbose = FALSE),
+      statistic = .boot_calculate_rmse,
+      R = iterations,
+      sim = "ordinary",
+      model = model,
+      normalized = normalized
+    )
+  }
+  result
+}