+++ tests/testthat/test-model_performance.bayesian.R
@@ -2,9 +2,9 @@ test_that("model_performance.stanreg", {
   skip_on_cran()
   skip_if_not_installed("curl")
   skip_if_offline()
-  skip_if_not_installed("httr")
+  skip_if_not_installed("httr2")
   set.seed(333)
-  model <- tryCatch(insight::download_model("stanreg_lm_1"), error = function(e) NULL)
+  model <- insight::download_model("stanreg_lm_1")
   skip_if(is.null(model))
   perf <- model_performance(model)
 
@@ -12,7 +12,7 @@ test_that("model_performance.stanreg", {
   expect_equal(perf$R2_adjusted, 0.7162912, tolerance = 1e-3)
   expect_equal(perf$ELPD, -83.49838, tolerance = 1e-3)
 
-  model <- tryCatch(insight::download_model("stanreg_lm_2"), error = function(e) NULL)
+  model <- insight::download_model("stanreg_lm_2")
   skip_if(is.null(model))
   perf <- model_performance(model)
 
@@ -20,7 +20,7 @@ test_that("model_performance.stanreg", {
   expect_equal(perf$R2_adjusted, 0.7979026, tolerance = 1e-3)
   expect_equal(perf$ELPD, -78.38735, tolerance = 1e-3)
 
-  model <- tryCatch(insight::download_model("stanreg_lmerMod_1"), error = function(e) NULL)
+  model <- insight::download_model("stanreg_lmerMod_1")
   skip_if(is.null(model))
   perf <- model_performance(model)
 
@@ -34,10 +34,10 @@ test_that("model_performance.brmsfit", {
   skip_on_cran()
   skip_if_not_installed("curl")
   skip_if_offline()
-  skip_if_not_installed("httr")
+  skip_if_not_installed("httr2")
   set.seed(333)
 
-  model <- tryCatch(insight::download_model("brms_1"), error = function(e) NULL)
+  model <- insight::download_model("brms_1")
   skip_if(is.null(model))
   expect_message({
     perf <- model_performance(model)
@@ -50,7 +50,7 @@ test_that("model_performance.brmsfit", {
     "RMSE", "Sigma"
   ))
 
-  model <- tryCatch(insight::download_model("brms_mixed_4"), error = function(e) NULL)
+  model <- insight::download_model("brms_mixed_4")
   skip_if(is.null(model))
   expect_message({
     perf <- model_performance(model)
@@ -63,7 +63,7 @@ test_that("model_performance.brmsfit", {
     "R2_adjusted", "R2_adjusted_marginal", "ICC", "RMSE", "Sigma"
   ))
 
-  model <- tryCatch(insight::download_model("brms_ordinal_1"), error = function(e) NULL)
+  model <- insight::download_model("brms_ordinal_1")
   skip_if(is.null(model))
   perf <- suppressWarnings(model_performance(model))
   expect_equal(perf$R2, 0.8760015, tolerance = 1e-3)